requirements.txt
alembic==1.4.3
certifi==2020.6.20
chardet==3.0.4
click==8.1.3
dnspython==2.0.0
email-validator==1.1.1
emoji==0.6.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.1.1
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.3
MarkupSafe==2.1.3
numpy==1.19.2
pandas==1.1.2
psycopg2==2.8.6
PyJWT==2.7.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
twilio==6.45.4
urllib3==1.25.10
Werkzeug==2.3.3
WTForms==2.3.3
apscheduler==3.10.1
